Refactor: useVirtualKeyboard 분리
